VPP-287: Remove logging in traps
Since Ruby 2.0, logging in a trap block is not allowed because the
default loggers all use mutexes to serialize the log writes and a mutex
is lock is not allowed in traps to prevent dead locks.  Removing the log
message prevents "log writing failed. can't be called from trap
context".