--thread-request
Allow the request payload to be put in a Thread-local variable, so that
client that are invoked during an RPC server request can use parts of
the request, most importantly, parts of "meta" in their own requests.
This is needed to pass along ownership information for db_connection in
Ecto 2.0.