Remove busy-wait
The `until quit` busy-wait loop consumes ~1 core for each instance of
carrot_rpc as the default `sleep 0` does the minimal amount of sleep
before waking up to check the boolean `quit`.  I've replaced it with an
IO.pipe and IO.select that does not consume any resources while it waits.
**NOTE: A Queue could not be used here because the MRI VM blocks use of
Mutexes inside signal handlers to prevent deadlocks because the Mutex
code is not-reentrant (i.e signal-interrupt-safe).  If a Queue is used
the thread silently fails with an exception and the signal is ignored.**

This is an incompatible changes because it removed the `-s`
(`--runloop_sleep`) option as it is no longer needed.